module3-task1: * move view templates
