+++ src/view/board/board.ts
